medical-data-analysis/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.335
langchain-openai==0.0.3
openai>=1.0.0
pydantic>=2.0.0

# Document processing
python-magic==0.4.27
beautifulsoup4==4.12.2
markdown==3.4.4
html2text==2020.1.16
pypdf2==3.0.1

# Data handling
numpy==1.24.3
pandas==2.0.3

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
colorlog==6.7.0

# Development dependencies
pytest==7.4.0
black==23.7.0
isort==5.12.0
mypy==1.5.1

# Data Extraction
tika==2.6.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==0.8.11

# Data Processing
scikit-learn==1.2.2
jsonpath-ng==1.5.0

# AI & NLP
transformers==4.28.1
torch==2.0.1
spacy==3.5.3
sentence-transformers==2.2.2

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
alembic==1.10.4

# API
fastapi==0.95.2
uvicorn==0.22.0
strawberry-graphql==0.175.1
redis==4.5.5

# Storage & Security
minio==7.1.15
cryptography==40.0.2
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Development & Testing
pytest-mock==3.11.1
pytest-cov==4.1.0
coverage==7.2.7
faker==18.13.0
freezegun==1.2.2
pdfminer.six>=20221105
pypdf>=3.0.0
lxml>=4.9.0
nltk>=3.7
scipy>=1.10.0
tiktoken>=0.4.0
llama-index>=0.8.0 